writeGraph
